pyproject.toml
[build-system]
requires = ["scikit-build-core>=0.9"]
build-backend = "scikit_build_core.build"

[project]
name = "neutts"
version = "1.2.0"
authors = [
  { name="neuphonic", email="general@neuphonic.com" },
]
description = "NeuTTS - a package for text-to-speech generation using Neuphonic's TTS models."
readme = "README.md"
requires-python = ">=3.10, <3.14"
dependencies = [
    "datasets>=4.6.1",
    "librosa==0.11.0",
    "neucodec>=0.0.4",
    "numpy~=2.2.6",
    "phonemizer>=3.0.0",
    "pyyaml>=6.0.3",
    "resemble-perth==1.0.1",
    "scikit-learn>=1.7.2",
    "scipy>=1.15.3",
    "soundfile==0.13.1",
    "tensorboard>=2.20.0",
    "torch>=2.8.0",
    "torchaudio>=2.10.0",
    "tqdm>=4.67.3",
    "transformers~=4.56.1",
]

# Defining optional installations
[project.optional-dependencies]
llama = ["llama-cpp-python"]
onnx = ["onnxruntime"]
all = ["llama-cpp-python", "onnxruntime"]

[tool.scikit-build]
# Include the espeak-ng-data directory that cmake copies into the package
wheel.packages = ["neutts", "neuttsair"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
